fix 64-bit DhbNumericalMethods failure
Patches applied:

 * bonzini@gnu.org--2004b/smalltalk--devo--2.2--patch-561
   fix 64-bit DhbNumericalMethods failure


git-archimport-id: bonzini@gnu.org--2004b/smalltalk--stable--2.3--patch-107

+++ libgst/interp.inl
@@ -86,7 +86,7 @@ mul_with_check (intptr_t a, intptr_t b)
   /* This fallback method uses a division to do overflow check */
   else
     {
-      if COMMON ((a | b) < (1L << (ST_INT_SIZE / 2))
+      if COMMON (((uintptr_t) (a | b)) < (1L << (ST_INT_SIZE / 2))
 	         || b == 0
 	         || result / b == a)
         return (result);
